+++ libs/balances-and-allowances/CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## [3.9.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.8.0...balances-and-allowances-v3.9.0) (2026-07-30)
+
+### Minor Changes
+
+- feat(balances): add hook to make token lists steady to prevent many requests to bw (#7887)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.11.0
+  - @cowprotocol/types@4.7.0
+  - @cowprotocol/wallet@3.10.0
+  - @cowprotocol/common-hooks@3.8.0
+  - @cowprotocol/core@3.6.4
+  - @cowprotocol/tokens@3.8.7
+  - @cowprotocol/common-const@3.7.1
+
+## [3.8.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.7.1...balances-and-allowances-v3.8.0) (2026-07-28)
+
+### Minor Changes
+
+- feat(bw): add sentry logging (#7873)
+
+- feat: add sentry logging for 429 error in balances multicall (#7868)
+
+- feat: use a single getPublicClient util (#7870)
+
+- feat(solana): load token balances (#7850)
+
+### Patch Changes
+
+- fix(bridge): fix bridge fee displaying (#7893)
+
+- fix(bridging): update bungee cctp to cctp-v2 (#7155)
+
+- Updated dependencies []:
+  - @cowprotocol/common-hooks@3.7.0
+  - @cowprotocol/types@4.6.0
+  - @cowprotocol/common-const@3.7.0
+  - @cowprotocol/common-utils@3.10.0
+  - @cowprotocol/core@3.6.3
+  - @cowprotocol/currency@1.2.2
+  - @cowprotocol/tokens@3.8.6
+  - @cowprotocol/wallet@3.9.0
+
+## [3.7.1](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.7.0...balances-and-allowances-v3.7.1) (2026-07-22)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.9.0
+  - @cowprotocol/common-hooks@3.6.1
+  - @cowprotocol/core@3.6.2
+  - @cowprotocol/tokens@3.8.5
+  - @cowprotocol/wallet@3.8.2
+
 ## [3.7.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.6.0...balances-and-allowances-v3.7.0) (2026-07-17)
 
 ### Minor Changes